# Optimize build/test/typecheck performance