--- a/.github/dependabot.yml
+++ b/.github/dependabot.yml
@@ -15,9 +15,6 @@ updates:
       - dependency-name: "@types/node"
         versions:
           - ">= 14.a, < 15"
-      - dependency-name: webpack-cli
-        versions:
-          - 4.5.0
       - dependency-name: "@babel/types"
         versions:
           - 7.12.12