--- a/.github/workflows/tests.yml
+++ b/.github/workflows/tests.yml
@@ -1,7 +1,9 @@
 name: Tests
 
 on:
-  push
+  push:
+  pull_request:
+    types: [review_requested]
 
 jobs:
   run-tests: