--- a/.gitignore
+++ b/.gitignore
@@ -5,6 +5,7 @@ node_modules
 .vscode-test-web/
 *.vsix
 .rubberduck/
+meta-*.json
 
 # Created by https://www.gitignore.io/api/node,visualstudiocode
 # Edit at https://www.gitignore.io/?templates=node,visualstudiocode