--- a/.vscode/tasks.json
+++ b/.vscode/tasks.json
@@ -8,7 +8,7 @@
       "script": "build",
       "group": "build",
       "isBackground": true,
-      "problemMatcher": ["$ts-webpack-watch"]
+      "problemMatcher": "$esbuild"
     }
   ]
 }