--- a/tsconfig.json
+++ b/tsconfig.json
@@ -18,7 +18,9 @@
     "noUnusedLocals": true,
     "resolveJsonModule": true,
     "skipLibCheck": true,
-    "isolatedModules": true
+    "isolatedModules": true,
+    "incremental": true,
+    "tsBuildInfoFile": "out/tsconfig.tsbuildinfo"
   },
   "exclude": [
     "node_modules",