As mentioned in https://github.com/nicoespeon/abracadabra/discussions/959
Now tsc is only used for typechecking.
The node/web extensions are bundled by esbuild with sourcemaps and bundle meta in development, minified in production.
`run-p` has been removed because of the overhead, it's faster to build both bundles in parallel in a single process (or even sequentially).
The entire src folder is also built with tsup/esbuild to be used by the contract tests. I haven't really found a way to improve the jest testing time.

`yarn build:node`: 73s -> 0.5s, `yarn build`: 0.8s
`yarn pretest:contract`: 26s -> 1s
`yarn typecheck` with incremental tsconfig: 24s -> 2s (3s if a single file was changed)