@nicoespeon 2023-12-29
@j4k0xb awesome! I'm gonna try these out and check that it doesn't cause a regression with the builds today. I'll let you know how it goes 😉