@nicoespeon 2023-12-29
It just works 🌈 
I just added some documentation and updated the yarn.lock file.

@j4k0xb thank you very, very much. This is one of the most impactful external contributions to the project I've seen in a while 🤩 

![gourmet](https://github.com/nicoespeon/abracadabra/assets/1094774/f697fb08-0586-4aa7-a626-323f8a14bd6c)